[Ref] Fix a ton of 64bit warnings in libopenmpt MSVC.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8708 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ sounddsp/Reverb.cpp
@@ -67,7 +67,7 @@ CReverb::CReverb()
 //----------------
 {
 	// Shared reverb state
-	InitMixBuffer(MixReverbBuffer, mpt::size(MixReverbBuffer));
+	InitMixBuffer(MixReverbBuffer, static_cast<uint32>(mpt::size(MixReverbBuffer)));
 	gnRvbROfsVol = 0;
 	gnRvbLOfsVol = 0;
